+++ mlrun/runtimes/nuclio.py
@@ -11,22 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import functools
 import inspect
 import json
 import socket
 
-import semver
-
-import mlrun.errors
-
-from ..config import config as mlconf
 from ..db import get_or_set_dburl
 from ..execution import MLClientCtx
 from ..model import RunTemplate
 from ..serving.server import v2_serving_init
 from ..serving.v1_serving import nuclio_serving_init
-from ..utils import logger
 from .local import get_func_arg
 from .serving import serving_subkind
 
@@ -90,43 +83,3 @@ def nuclio_jobs_handler(context, event):
         err = str(exc)
         ctx.set_state(error=err)
     return ctx.to_json()
-
-
-def validate_version_compatibility(min_nuclio_version):
-    """
-    Validation is best effort - if we can't parse we assume compatible.
-    """
-    try:
-        parsed_current_version = semver.VersionInfo.parse(mlconf.nuclio_version)
-        parsed_min_version = semver.VersionInfo.parse(min_nuclio_version)
-    except ValueError:
-        logger.warning(
-            "Unable to parse nuclio version, assuming compatibility",
-            installed_nuclio_version=mlconf.nuclio_version,
-            min_nuclio_version=min_nuclio_version,
-        )
-        return True
-
-    if parsed_current_version >= parsed_min_version:
-        return True
-    return False
-
-
-def min_versions(*versions):
-    def decorator(function):
-        # functools.wraps copies the metadata of function
-        @functools.wraps(function)
-        def wrapper(*args, **kwargs):
-            for version in versions:
-                if validate_version_compatibility(version):
-                    return function(*args, **kwargs)
-
-            message = (
-                f"{function.__name__} is supported since nuclio {' or '.join(versions)}, currently using "
-                f"nuclio {mlconf.nuclio_version}, please upgrade."
-            )
-            raise mlrun.errors.MLRunMissingDependencyError(message)
-
-        return wrapper
-
-    return decorator
